ai-agent/requirements.txt
# --- Core Web Framework ---
fastapi
uvicorn[standard]
# --- AI & LLM Services ---
groq
pinecone-client
# --- LangChain Components (Modern) ---
langchain
langchain-groq
langchain-community
langchain-huggingface  # New package for HuggingFace embeddings
langchain-pinecone      # New package for Pinecone integration
langchain-chroma        # New package for Chroma integration
# --- Vector Stores & Data Handling ---
chromadb
PyMuPDF
pypdf
# --- Core ML & Data Science ---
sentence-transformers
huggingface-hub
torch
transformers
scikit-learn
pandas
numpy
# --- Utilities ---
python-dotenv